Code refactoring, db metadata rendering moved to a single file
