+++ src/renderer/containers/app.css
@@ -34,6 +34,9 @@ span.list-title {
 }
 
 span.list-item {
-  cursor: pointer;
   word-break: break-all;
+}
+
+span.clickable {
+  cursor: pointer;
 }
